restructure pages to fix build errors
