+++ web/src/pages/private-office/index.js
@@ -1,14 +1,14 @@
 import React from "react";
-import WorkspacePageShell from "../workspace/shared/WorkspacePageShell";
-import WorkspaceSparkHero from "../workspace/shared/WorkspaceSparkHero";
-import WorkspaceSparkSection from "../workspace/shared/WorkspaceSparkSection";
-import WorkspaceSparkAmenities from "../workspace/shared/WorkspaceSparkAmenities";
-import WorkspaceSparkContact from "../workspace/shared/WorkspaceSparkContact";
-import WorkspaceSparkCta from "../workspace/shared/WorkspaceSparkCta";
-import { WORKSPACE_CTAS } from "../workspace/constants/ctaPlaceholders";
-import { DUMMY_PRIVATE_OFFICES } from "../workspace/data/dummyPrivateOffices";
+import WorkspacePageShell from "../../workspace/shared/WorkspacePageShell";
+import WorkspaceSparkHero from "../../workspace/shared/WorkspaceSparkHero";
+import WorkspaceSparkSection from "../../workspace/shared/WorkspaceSparkSection";
+import WorkspaceSparkAmenities from "../../workspace/shared/WorkspaceSparkAmenities";
+import WorkspaceSparkContact from "../../workspace/shared/WorkspaceSparkContact";
+import WorkspaceSparkCta from "../../workspace/shared/WorkspaceSparkCta";
+import { WORKSPACE_CTAS } from "../../workspace/constants/ctaPlaceholders";
+import { DUMMY_PRIVATE_OFFICES } from "../../workspace/data/dummyPrivateOffices";
 import privateOfficeImage from "../../images/workspace/offering-private-office.jpg";
-import * as styles from "../workspace/shared/workspaceShared.module.scss";
+import * as styles from "../../workspace/shared/workspaceShared.module.scss";
 
 const PrivateOfficePage = () => (
   <WorkspacePageShell>
