+++ web/src/workspace/constants/ctaPlaceholders.js
