+++ web/src/workspace/shared/RoomCapacityIcons.js
