+++ web/src/workspace/shared/WorkspaceMembershipBand.js
