+++ web/src/workspace/shared/WorkspaceSparkHero.js
